separate code according to purpose using cogs
